int => float

+++ fonction.py
@@ -1,7 +1,7 @@
-Ya = int(input("Entrez l'ordonnées du point A "))
-Xa = int(input("Entrez l'absice du point A "))
-Yb = int(input("Entrez l'ordonnées du point B "))
-Xb = int(input("Entrez l'absice du point B ")) 
+Ya = float(input("Entrez l'ordonnées du point A "))
+Xa = float(input("Entrez l'absice du point A "))
+Yb = float(input("Entrez l'ordonnées du point B "))
+Xb = float(input("Entrez l'absice du point B ")) 
 
 coef_dir = (Ya - Yb) / (Xa - Xb)
 ordo_origine = Ya - coef_dir * Xa
